Fix webapp and actuator security for the admin webapp
